+++ hosts/2600AD/hardware-configuration.nix
@@ -13,6 +13,8 @@
 
   # ZFS filesystem support
   boot.supportedFilesystems = [ "zfs" ];
+  boot.zfs.forceImportAll = true; # remove this after we've fixed our networking/hostid shennanigans
+  boot.zfs.forceImportRoot = true; # requred by above, also remove after network is un-borked
 
   # ─────────────────────────────────────────────────────────────────────────
   # LUKS Encrypted Devices (Atari-themed labels)